Remove shared mockdata
