+++ cms/src/components/content/EditContentValidation.spec.ts
@@ -1,7 +1,7 @@
 import "fake-indexeddb/auto";
 import { describe, it, afterEach, expect } from "vitest";
 import { mount } from "@vue/test-utils";
-import * as mockData from "@/mockdata";
+import * as mockData from "@/tests/mockdata";
 import EditContentValidation from "./EditContentValidation.vue";
 
 describe("EditContentValidation.vue", () => {
