+++ cms/src/components/content/TagSelector.spec.ts
@@ -5,7 +5,7 @@ import { createTestingPinia } from "@pinia/testing";
 import { setActivePinia } from "pinia";
 import TagSelector from "./TagSelector.vue";
 import { db, TagType, type ContentDto, accessMap } from "luminary-shared";
-import * as mockData from "@/mockdata";
+import * as mockData from "@/tests/mockdata";
 import waitForExpect from "wait-for-expect";
 import { Combobox } from "@headlessui/vue";
