--- shared/src/mockdata
@@ -1,1 +1,0 @@
-../../mockdata
