+++ shared/src/tests/mockdata.ts
@@ -1,0 +1,369 @@
+import {
+    DocType,
+    type ContentDto,
+    type PostDto,
+    type LocalChangeDto,
+    type LanguageDto,
+    type TagDto,
+    // AclPermission,
+    type ImageDto,
+    // GroupDto,
+    TagType,
+    PublishStatus,
+} from "../types";
+
+export const mockCategoryDto: TagDto = {
+    _id: "tag-category1",
+    type: DocType.Tag,
+    updatedTimeUtc: 1704114000000,
+    memberOf: ["group-public-content"],
+    tagType: TagType.Category,
+    pinned: false,
+    image: "test-image.jpg",
+    tags: [],
+};
+
+export const mockTopicDto: TagDto = {
+    _id: "tag-topicA",
+    type: DocType.Tag,
+    updatedTimeUtc: 1704114000000,
+    memberOf: ["group-public-content"],
+    tagType: TagType.Category,
+    pinned: false,
+    image: "test-image.jpg",
+    tags: [],
+};
+
+export const mockPostDto: PostDto = {
+    _id: "post-post1",
+    type: DocType.Post,
+    updatedTimeUtc: 1704114000000,
+    memberOf: ["group-public-content"],
+    image: "test-image.jpg",
+    tags: ["tag-category1"],
+};
+
+export const mockEnglishContentDto: ContentDto = {
+    _id: "content-post1-eng",
+    type: DocType.Content,
+    parentId: "post-post1",
+    parentType: DocType.Post,
+    updatedTimeUtc: 1704114000000,
+    memberOf: ["group-public-content"],
+    tags: ["tag-category1"],
+    language: "lang-eng",
+    status: PublishStatus.Published,
+    slug: "post1-eng",
+    title: "Post 1",
+    summary: "This is an example post",
+    author: "ChatGPT",
+    text: '{"type":"doc","content":[{"type":"paragraph","content":[{"type":"text","text":"In the quiet town of Willowdale, little Lily wept as her beloved cat, Whiskers, went missing. Frantically searching the neighborhood, she stumbled upon Fireman Jake, known for his kind heart. With a reassuring smile, he promised to help. Lily clung to hope as they combed the streets together. Beneath a dusty porch, they found Whiskers, scared but unharmed. Grateful tears filled Lily\'s eyes as Fireman Jake handed her the rescued feline. Their small town echoed with cheers as Lily hugged her furry friend, and from that day forward, Fireman Jake became a hero in her heart and the community\'s beloved guardian"}]}]}',
+    localisedImage: "",
+    audio: "",
+    video: "",
+    publishDate: 1704114000000,
+    expiryDate: undefined,
+    image: "test-image.jpg",
+};
+export const mockFrenchContentDto: ContentDto = {
+    _id: "content-post1-fra",
+    type: DocType.Content,
+    parentId: "post-post1",
+    parentType: DocType.Post,
+    updatedTimeUtc: 1704114000000,
+    memberOf: ["group-public-content"],
+    tags: ["tag-category1"],
+    language: "lang-fra",
+    status: PublishStatus.Published,
+    slug: "post1-fra",
+    title: "Post 1",
+    summary: "Ceci est un exemple de publication.",
+    author: "ChatGPT",
+    text: '{"type":"doc","content":[{"type":"paragraph","content":[{"type":"text","text":"Dans la paisible ville de Willowdale, la petite Lily pleurait la disparition de son cher chat, Whiskers. Cherchant frénétiquement dans le quartier, elle tomba sur le pompier Jake, réputé pour son cœur généreux. Avec un sourire rassurant, il promit de l\'aider. Lily s\'accrocha à l\'espoir alors qu\'ils parcouraient les rues ensemble. Sous un porche poussiéreux, ils trouvèrent Whiskers, effrayé mais sain et sauf. Des larmes de gratitude remplirent les yeux de Lily lorsque le pompier Jake lui remit le félin sauvé. Leur petite ville résonna de joie tandis que Lily serrait son ami à fourrure dans ses bras, et dès ce jour, le pompier Jake devint un héros dans son cœur et le gardien bien-aimé de la communauté."}]}]}',
+    localisedImage: "",
+    audio: "",
+    video: "",
+    publishDate: 1704114000000,
+    expiryDate: undefined,
+    image: "test-image.jpg",
+};
+export const mockSwahiliContentDto: ContentDto = {
+    _id: "content-post1-swa",
+    type: DocType.Content,
+    parentId: "post-post1",
+    parentType: DocType.Post,
+    updatedTimeUtc: 1704114000000,
+    memberOf: ["group-public-content"],
+    tags: ["tag-category1"],
+    language: "lang-swa",
+    status: PublishStatus.Published,
+    slug: "post1-swa",
+    title: "Post 1",
+    summary: "Hii ni chapisho la mfano.",
+    author: "ChatGPT",
+    text: '{"type":"doc","content":[{"type":"paragraph","content":[{"type":"text","text":"Dans la paisible ville de Willowdale, la petite Lily pleurait la disparition de son cher chat, Whiskers. Cherchant frénétiquement dans le quartier, elle tomba sur le pompier Jake, réputé pour son cœur généreux. Avec un sourire rassurant, il promit de l\'aider. Lily s\'accrocha à l\'espoir alors qu\'ils parcouraient les rues ensemble. Sous un porche poussiéreux, ils trouvèrent Whiskers, effrayé mais sain et sauf. Des larmes de gratitude remplirent les yeux de Lily lorsque le pompier Jake lui remit le félin sauvé. Leur petite ville résonna de joie tandis que Lily serrait son ami à fourrure dans ses bras, et dès ce jour, le pompier Jake devint un héros dans son cœur et le gardien bien-aimé de la communauté."}]}]}',
+    localisedImage: "",
+    audio: "",
+    video: "",
+    publishDate: 1704114000000,
+    expiryDate: undefined,
+    image: "test-image.jpg",
+};
+export const mockCategoryContentDto: ContentDto = {
+    _id: "content-tag-category1",
+    type: DocType.Content,
+    parentId: "tag-category1",
+    parentType: DocType.Tag,
+    updatedTimeUtc: 1704114000000,
+    memberOf: [],
+    tags: [],
+    language: "lang-eng",
+    status: PublishStatus.Published,
+    slug: "content-tag-category1",
+    title: "Category 1",
+    summary: "Example tag",
+    text: '{"type":"doc","content":[{"type":"paragraph","content":[{"type":"text","text":"A category"}]}]}',
+    publishDate: 1704114000000,
+    image: "test-image.jpg",
+};
+
+export const mockTopicContentDto: ContentDto = {
+    _id: "content-tag-topicA",
+    type: DocType.Content,
+    parentId: "tag-topicA",
+    parentType: DocType.Tag,
+    updatedTimeUtc: 1704114000000,
+    memberOf: [],
+    tags: [],
+    language: "lang-eng",
+    status: PublishStatus.Published,
+    slug: "content-tag-topicA",
+    title: "Topic A",
+    summary: "Example tag",
+    text: '{"type":"doc","content":[{"type":"paragraph","content":[{"type":"text","text":"A category"}]}]}',
+    publishDate: 1704114000000,
+    image: "test-image.jpg",
+};
+
+export const mockLanguageDtoEng: LanguageDto = {
+    _id: "lang-eng",
+    type: DocType.Language,
+    updatedTimeUtc: 1704114000000,
+    memberOf: ["group-languages"],
+    languageCode: "eng",
+    name: "English",
+};
+export const mockLanguageDtoFra: LanguageDto = {
+    _id: "lang-fra",
+    type: DocType.Language,
+    updatedTimeUtc: 1704114000000,
+    memberOf: ["group-languages"],
+    languageCode: "fra",
+    name: "Français",
+};
+
+export const mockLanguageDtoSwa: LanguageDto = {
+    _id: "lang-swa",
+    type: DocType.Language,
+    updatedTimeUtc: 1704114000000,
+    memberOf: ["group-languages"],
+    languageCode: "swa",
+    name: "Swahili",
+};
+
+export const mockImageDto: ImageDto = {
+    _id: "image-image1",
+    type: DocType.Image,
+    name: "Image 1",
+    description: "An image",
+    updatedTimeUtc: Date.now(),
+    memberOf: ["group-private-content"],
+    fileCollections: [
+        {
+            aspectRatio: 1.5,
+            imageFiles: [
+                {
+                    width: 180,
+                    height: 120,
+                    filename: "test-image.webp",
+                },
+            ],
+        },
+    ],
+};
+
+export const mockLocalChange1: LocalChangeDto = {
+    id: 42,
+    doc: mockPostDto,
+    docId: "post-post1",
+};
+export const mockLocalChange2: LocalChangeDto = {
+    id: 43,
+    doc: mockEnglishContentDto,
+    docId: "content-post1-eng",
+};
+
+export const fullAccessToAllContentMap = {
+    "group-private-content": {
+        post: {
+            view: true,
+            create: true,
+            edit: true,
+            translate: true,
+            publish: true,
+        },
+        tag: {
+            view: true,
+            create: true,
+            edit: true,
+            translate: true,
+            publish: true,
+            assign: true,
+        },
+        language: {
+            view: true,
+            create: true,
+            edit: true,
+            translate: true,
+            publish: true,
+        },
+    },
+    "group-public-content": {
+        post: {
+            view: true,
+            create: true,
+            edit: true,
+            translate: true,
+            publish: true,
+        },
+        tag: {
+            view: true,
+            create: true,
+            edit: true,
+            translate: true,
+            publish: true,
+            assign: true,
+        },
+        language: {
+            view: true,
+            create: true,
+            edit: true,
+            translate: true,
+            publish: true,
+        },
+    },
+    "group-languages": {
+        post: {
+            view: true,
+            create: true,
+            edit: true,
+            translate: true,
+            publish: true,
+        },
+        tag: {
+            view: true,
+            create: true,
+            edit: true,
+            translate: true,
+            publish: true,
+            assign: true,
+        },
+        language: {
+            view: true,
+            create: true,
+            edit: true,
+            translate: true,
+            publish: true,
+        },
+    },
+};
+
+export const viewAccessToAllContentMap = {
+    "group-private-content": {
+        post: {
+            view: true,
+            create: false,
+            edit: false,
+            translate: false,
+            publish: false,
+        },
+        tag: {
+            view: true,
+            create: false,
+            edit: false,
+            translate: false,
+            publish: false,
+        },
+        language: {
+            view: true,
+            create: false,
+            edit: false,
+            translate: false,
+            publish: false,
+        },
+    },
+    "group-public-content": {
+        post: {
+            view: true,
+            create: false,
+            edit: false,
+            translate: false,
+            publish: false,
+        },
+        tag: {
+            view: true,
+            create: false,
+            edit: false,
+            translate: false,
+            publish: false,
+        },
+        language: {
+            view: true,
+            create: false,
+            edit: false,
+            translate: false,
+            publish: false,
+        },
+    },
+    "group-languages": {
+        post: {
+            view: true,
+            create: false,
+            edit: false,
+            translate: false,
+            publish: false,
+        },
+        tag: {
+            view: true,
+            create: false,
+            edit: false,
+            translate: false,
+            publish: false,
+        },
+        language: {
+            view: true,
+            create: false,
+            edit: false,
+            translate: false,
+            publish: false,
+        },
+    },
+};
+
+export const translateAccessToAllContent = {
+    "group-private-content": {
+        post: { view: true, translate: true },
+        tag: { view: true, translate: true },
+        language: { view: true, translate: true },
+    },
+    "group-public-content": {
+        post: { view: true, translate: true },
+        tag: { view: true, translate: true },
+        language: { view: true, translate: true },
+    },
+    "group-languages": {
+        post: { view: true, translate: true },
+        tag: { view: true, translate: true },
+        language: { view: true, translate: true },
+    },
+};
